refactor modules::types
