+++ src/modules/mod.rs
@@ -2,5 +2,6 @@
 mod translator;
 pub use translator::Translator;
 
+pub mod types;
 pub mod net;
 pub use net::{ClientController, ConnectionManager};
