+++ src/modules/net/types.rs
@@ -1,9 +1,7 @@
 use serde::{Deserialize, Serialize};
-use serde_json::Value;
 
+use super::PlayerId;
 use crate::generic::ReactorID;
-
-pub type PlayerId = u64;
 
 #[derive(Serialize, Deserialize, Clone, Key, Debug)]
 pub struct Register {
@@ -18,15 +16,4 @@ pub struct Accepted {
 }
 
 #[derive(Serialize, Deserialize, Clone, Key, Debug)]
-pub struct Data {
-    pub value: Value,
-}
-
-#[derive(Serialize, Deserialize, Clone, Key, Debug)]
 pub struct Close {}
-
-#[derive(Serialize, Deserialize, Clone, Key, Debug)]
-pub struct PlayerMsg {
-    pub value: Value,
-    pub id: PlayerId,
-}
